requirements.txt
pytorch-lightning==1.2.3
torch==1.9.1
wandb==0.10.18
plotly==4.14.3
scikit-learn==0.24.1
joblib~=1.0.0
hydra-core~=1.0.5
pandas==1.2.1
pygtrie==2.4.2
seaborn==0.11.1
dgl-cu110==0.6.1
WTForms==2.3.3
flask_wtf==0.15.1
email_validator==1.1.3
python-decouple==3.4
jinja2==3.0.1
pymongo
fastapi
uvicorn
pygtrie
